Move essential target classes to target plugins
