+++ base/uk.ac.stfc.isis.ibex.synoptic/META-INF/MANIFEST.MF
@@ -24,3 +24,4 @@ Export-Package: uk.ac.stfc.isis.ibex.synoptic,
  uk.ac.stfc.isis.ibex.synoptic.model.targets,
  uk.ac.stfc.isis.ibex.synoptic.navigation,
  uk.ac.stfc.isis.ibex.synoptic.xml
+Import-Package: uk.ac.stfc.isis.ibex.targets
