+++ base/uk.ac.stfc.isis.ibex.synoptic/src/uk/ac/stfc/isis/ibex/synoptic/model/Component.java
@@ -22,6 +22,8 @@ package uk.ac.stfc.isis.ibex.synoptic.model;
 import java.util.List;
 import java.util.Set;
 
+import uk.ac.stfc.isis.ibex.targets.Target;
+
 public interface Component {
 	String name();
