+++ base/uk.ac.stfc.isis.ibex.synoptic/src/uk/ac/stfc/isis/ibex/synoptic/model/ComponentType.java
@@ -24,7 +24,8 @@ import java.util.Collections;
 import java.util.List;
 
 import uk.ac.stfc.isis.ibex.synoptic.model.targets.PerspectiveTarget;
-import uk.ac.stfc.isis.ibex.synoptic.model.targets.ViewTarget;
+import uk.ac.stfc.isis.ibex.targets.Target;
+import uk.ac.stfc.isis.ibex.targets.ViewTarget;
 
 /**
  * The type of components available to the synoptic.
