+++ base/uk.ac.stfc.isis.ibex.synoptic/src/uk/ac/stfc/isis/ibex/synoptic/model/targets/PerspectiveTarget.java
@@ -19,7 +19,7 @@
 
 package uk.ac.stfc.isis.ibex.synoptic.model.targets;
 
-import uk.ac.stfc.isis.ibex.synoptic.model.Target;
+import uk.ac.stfc.isis.ibex.targets.Target;
 
 /*
  * Switch to a different Isis perspective from the synoptic
