+++ base/uk.ac.stfc.isis.ibex.synoptic/src/uk/ac/stfc/isis/ibex/synoptic/model/targets/TargetBuilder.java
@@ -22,9 +22,9 @@ package uk.ac.stfc.isis.ibex.synoptic.model.targets;
 import java.util.Arrays;
 
 import uk.ac.stfc.isis.ibex.synoptic.model.Component;
-import uk.ac.stfc.isis.ibex.synoptic.model.Target;
 import uk.ac.stfc.isis.ibex.synoptic.model.desc.Property;
 import uk.ac.stfc.isis.ibex.synoptic.model.desc.TargetDescription;
+import uk.ac.stfc.isis.ibex.targets.Target;
 
 public class TargetBuilder {
