+++ base/uk.ac.stfc.isis.ibex.synoptic/src/uk/ac/stfc/isis/ibex/synoptic/navigation/InstrumentNavigationGraph.java
@@ -29,8 +29,8 @@ import com.google.common.base.Strings;
 
 import uk.ac.stfc.isis.ibex.synoptic.model.Component;
 import uk.ac.stfc.isis.ibex.synoptic.model.Synoptic;
-import uk.ac.stfc.isis.ibex.synoptic.model.Target;
 import uk.ac.stfc.isis.ibex.synoptic.model.targets.GroupedComponentTarget;
+import uk.ac.stfc.isis.ibex.targets.Target;
 
 public class InstrumentNavigationGraph {
