+++ base/uk.ac.stfc.isis.ibex.targets/META-INF/MANIFEST.MF
@@ -8,3 +8,4 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ActivationPolicy: lazy
+Export-Package: uk.ac.stfc.isis.ibex.targets
