+++ base/uk.ac.stfc.isis.ibex.ui.synoptic/src/uk/ac/stfc/isis/ibex/ui/synoptic/SynopticPresenter.java
@@ -44,17 +44,17 @@ import uk.ac.stfc.isis.ibex.synoptic.ObservingSynopticModel;
 import uk.ac.stfc.isis.ibex.synoptic.Synoptic;
 import uk.ac.stfc.isis.ibex.synoptic.SynopticModel;
 import uk.ac.stfc.isis.ibex.synoptic.model.Component;
-import uk.ac.stfc.isis.ibex.synoptic.model.Target;
 import uk.ac.stfc.isis.ibex.synoptic.model.desc.SynopticDescription;
 import uk.ac.stfc.isis.ibex.synoptic.model.targets.GroupedComponentTarget;
 import uk.ac.stfc.isis.ibex.synoptic.model.targets.OpiTarget;
 import uk.ac.stfc.isis.ibex.synoptic.model.targets.PerspectiveTarget;
-import uk.ac.stfc.isis.ibex.synoptic.model.targets.ViewTarget;
 import uk.ac.stfc.isis.ibex.synoptic.navigation.TargetNode;
+import uk.ac.stfc.isis.ibex.targets.Target;
+import uk.ac.stfc.isis.ibex.targets.ViewTarget;
 import uk.ac.stfc.isis.ibex.ui.UI;
 import uk.ac.stfc.isis.ibex.ui.synoptic.views.LinkedViews;
-import uk.ac.stfc.isis.ibex.ui.synoptic.views.OpiTargetView;
 import uk.ac.stfc.isis.ibex.ui.synoptic.views.SynopticView;
+import uk.ac.stfc.isis.ibex.ui.targets.OpiTargetView;
 
 /**
  * Responsible for the presentation logic of the synoptic.
